transfroming coordinates moved from RotateObject3D.tsx to object CoordinateTransformer. Forgot to merge sideBar last time - commited now.
