+++ package.json
@@ -7,11 +7,13 @@
     "dev": "vite",
     "build": "tsc && vite build",
     "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
-    "preview": "vite preview"
+    "preview": "vite preview",
+    "test": "vitest"
   },
   "dependencies": {
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "react-scrollbars-custom": "4.1.1",
     "styled-components": "6.1.11",
     "three": "0.162.0"
   },
